+++ crates/rslint_core/CHANGELOG.md
@@ -16,6 +16,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Added benchmarks for linting a file
 - Added command descriptors
 - Added `File` from rslint_cli
+- Added the `regex` group
+- Added `no-invalid-regexp`
 
 ### Changed